refactor: extract IdentityFields from ReIdentifyModal (#520)
Pulls the TMDB search box, title input, media-type toggle, and season
input out of ReIdentifyModal into a controlled IdentityFields component
so the upcoming arm-a-drive modal can reuse the same form without a
job object. Job-specific chrome (header, wrong-ID notice, same-name
quick-pick, action buttons, status bar) stays in ReIdentifyModal.
