+++ frontend/src/components/ReIdentifyModal.tsx
@@ -1,17 +1,9 @@
-import { useState, useRef, useEffect, useCallback, useMemo, type KeyboardEvent } from 'react';
-import { motion, AnimatePresence } from 'motion/react';
-import { IcoDisc, IcoMovie, IcoTv, IcoSearch, IcoRetry } from '../app/components/icons';
+import { useState, useMemo, type KeyboardEvent } from 'react';
+import { motion } from 'motion/react';
+import { IcoDisc, IcoTv, IcoRetry } from '../app/components/icons';
 import type { Job } from '../types';
 import { SvPanel, SvLabel, sv } from '../app/components/synapse';
-
-interface TmdbResult {
-    tmdb_id: number;
-    name: string;
-    type: 'tv' | 'movie';
-    year: string;
-    poster_path: string | null;
-    popularity: number;
-}
+import IdentityFields, { type IdentityValue } from './IdentityFields';
 
 /** A same-name TMDB candidate persisted on the job at identify time. */
 interface Candidate {
@@ -28,63 +20,16 @@ interface ReIdentifyModalProps {
 }
 
 export default function ReIdentifyModal({ job, onSubmit, onCancel }: ReIdentifyModalProps) {
-    const [title, setTitle] = useState(job.detected_title || '');
-    const [contentType, setContentType] = useState<'movie' | 'tv'>(
-        job.content_type === 'tv' ? 'tv' : 'movie'
-    );
-    const [season, setSeason] = useState<string>(String(job.detected_season || 1));
-    const [tmdbId, setTmdbId] = useState<number | undefined>();
-    // First-air year of the selected TMDB result, retained alongside tmdbId so the
-    // confirmation line can show "Name (year) · TMDB #id". Cleared when the user
-    // types a manual title (no confirmed match anymore).
-    const [selectedYear, setSelectedYear] = useState<string | undefined>();
-    const [searchQuery, setSearchQuery] = useState('');
-    const [searchResults, setSearchResults] = useState<TmdbResult[]>([]);
-    const [isSearching, setIsSearching] = useState(false);
-    const titleInputRef = useRef<HTMLInputElement>(null);
-    const searchTimerRef = useRef<ReturnType<typeof setTimeout>>();
-
-    useEffect(() => {
-        titleInputRef.current?.focus();
-    }, []);
-
-    useEffect(() => () => {
-        if (searchTimerRef.current) clearTimeout(searchTimerRef.current);
-    }, []);
-
-    const doSearch = useCallback(async (query: string) => {
-        if (!query.trim()) {
-            setSearchResults([]);
-            return;
-        }
-        setIsSearching(true);
-        try {
-            const resp = await fetch(`/api/tmdb/search?query=${encodeURIComponent(query)}`);
-            if (resp.ok) {
-                const data = await resp.json();
-                setSearchResults(data.results || []);
-            }
-        } catch {
-            // Silently fail — search is optional
-        } finally {
-            setIsSearching(false);
-        }
-    }, []);
-
-    const handleSearchChange = (value: string) => {
-        setSearchQuery(value);
-        if (searchTimerRef.current) clearTimeout(searchTimerRef.current);
-        searchTimerRef.current = setTimeout(() => doSearch(value), 500);
-    };
-
-    const selectResult = (result: TmdbResult) => {
-        setTitle(result.name);
-        setContentType(result.type);
-        setTmdbId(result.tmdb_id);
-        setSelectedYear(result.year || undefined);
-        setSearchResults([]);
-        setSearchQuery('');
-    };
+    const [identity, setIdentity] = useState<IdentityValue>({
+        title: job.detected_title || '',
+        contentType: job.content_type === 'tv' ? 'tv' : 'movie',
+        season: String(job.detected_season || 1),
+        tmdbId: undefined,
+    });
+    // Mirrors IdentityFields' internal search-query state so the Enter-to-submit
+    // guard below can still avoid submitting while the user is mid-search (the
+    // search box itself is owned by IdentityFields now).
+    const [hasSearchQuery, setHasSearchQuery] = useState(false);
 
     // Same-name twins recorded at identify time (e.g. Frasier 1993 + 2023). When
     // present, they drive a one-click "Did you mean?" picker so the user skips the
@@ -117,32 +62,19 @@ export default function ReIdentifyModal({ job, onSubmit, onCancel }: ReIdentifyM
     };
 
     const handleSubmit = () => {
-        if (!title.trim()) return;
+        if (!identity.title.trim()) return;
         onSubmit(
-            title.trim(),
-            contentType,
-            contentType === 'tv' ? (parseInt(season, 10) || 1) : undefined,
-            tmdbId,
+            identity.title.trim(),
+            identity.contentType,
+            identity.contentType === 'tv' ? (parseInt(identity.season, 10) || 1) : undefined,
+            identity.tmdbId,
         );
     };
 
     const handleKeyDown = (e: KeyboardEvent) => {
-        if (e.key === 'Enter' && !searchQuery) handleSubmit();
+        if (e.key === 'Enter' && !hasSearchQuery) handleSubmit();
         if (e.key === 'Escape') onCancel();
     };
-
-    const inputStyle = (filled: boolean): React.CSSProperties => ({
-        width: '100%',
-        background: sv.bg1,
-        border: `1px solid ${filled ? sv.lineHi : sv.lineMid}`,
-        color: sv.cyanHi,
-        fontFamily: sv.mono,
-        fontSize: 13,
-        padding: '10px 12px',
-        outline: 'none',
-        boxShadow: filled ? `0 0 12px ${sv.cyan}33, inset 0 0 8px ${sv.cyan}0d` : 'none',
-        transition: 'border-color 0.18s',
-    });
 
     return (
         <motion.div
@@ -272,7 +204,7 @@ export default function ReIdentifyModal({ job, onSubmit, onCancel }: ReIdentifyM
                                     a TMDB id is committed (an ambiguous disc has none)
                                     and the user hasn't yet picked a replacement, so they
                                     can compare wrong-vs-right before re-identifying. */}
-                                {job.tmdb_id != null && tmdbId == null && (
+                                {job.tmdb_id != null && identity.tmdbId == null && (
                                     <p
                                         style={{
                                             fontFamily: sv.mono,
@@ -355,258 +287,12 @@ export default function ReIdentifyModal({ job, onSubmit, onCancel }: ReIdentifyM
 
                         <div style={{ height: 1, background: sv.line }} />
 
-                        {/* TMDB Search */}
-                        <div style={{ display: 'flex', flexDirection: 'column', gap: 8 }}>
-                            <SvLabel size={10}>Search TMDB</SvLabel>
-                            <div style={{ position: 'relative' }}>
-                                <IcoSearch
-                                    size={14}
-                                    color={sv.inkFaint}
-                                    style={{ position: 'absolute', left: 12, top: '50%', transform: 'translateY(-50%)' }}
-                                />
-                                <input
-                                    type="text"
-                                    value={searchQuery}
-                                    onChange={(e) => handleSearchChange(e.target.value)}
-                                    placeholder="Search for correct title..."
-                                    style={{ ...inputStyle(!!searchQuery), paddingLeft: 36 }}
-                                    onFocus={(e) => (e.currentTarget.style.borderColor = sv.cyan)}
-                                    onBlur={(e) =>
-                                        (e.currentTarget.style.borderColor = searchQuery ? sv.lineHi : sv.lineMid)
-                                    }
-                                />
-                                {isSearching && (
-                                    <motion.div
-                                        animate={{ rotate: 360 }}
-                                        transition={{ duration: 1, repeat: Infinity, ease: 'linear' }}
-                                        style={{ position: 'absolute', right: 12, top: '50%', transform: 'translateY(-50%)' }}
-                                    >
-                                        <IcoRetry size={14} color={sv.cyan} />
-                                    </motion.div>
-                                )}
-                            </div>
-
-                            <AnimatePresence>
-                                {searchResults.length > 0 && (
-                                    <motion.div
-                                        initial={{ opacity: 0, height: 0 }}
-                                        animate={{ opacity: 1, height: 'auto' }}
-                                        exit={{ opacity: 0, height: 0 }}
-                                        style={{
-                                            maxHeight: 192,
-                                            overflowY: 'auto',
-                                            border: `1px solid ${sv.line}`,
-                                            background: `${sv.bg1}80`,
-                                        }}
-                                    >
-                                        {searchResults.map((result) => (
-                                            <button
-                                                key={`${result.type}-${result.tmdb_id}`}
-                                                onClick={() => selectResult(result)}
-                                                style={{
-                                                    width: '100%',
-                                                    display: 'flex',
-                                                    alignItems: 'center',
-                                                    gap: 12,
-                                                    padding: '8px 12px',
-                                                    borderBottom: `1px solid ${sv.line}`,
-                                                    background: 'transparent',
-                                                    cursor: 'pointer',
-                                                    textAlign: 'left',
-                                                    transition: 'background 0.18s',
-                                                }}
-                                                onMouseEnter={(e) =>
-                                                    (e.currentTarget.style.background = `${sv.cyan}1a`)
-                                                }
-                                                onMouseLeave={(e) =>
-                                                    (e.currentTarget.style.background = 'transparent')
-                                                }
-                                            >
-                                                {result.poster_path ? (
-                                                    <img
-                                                        src={`https://image.tmdb.org/t/p/w92${result.poster_path}`}
-                                                        alt=""
-                                                        style={{
-                                                            width: 32,
-                                                            height: 48,
-                                                            objectFit: 'cover',
-                                                            flexShrink: 0,
-                                                            border: `1px solid ${sv.line}`,
-                                                        }}
-                                                    />
-                                                ) : (
-                                                    <div
-                                                        style={{
-                                                            width: 32,
-                                                            height: 48,
-                                                            background: sv.bg2,
-                                                            border: `1px solid ${sv.line}`,
-                                                            display: 'flex',
-                                                            alignItems: 'center',
-                                                            justifyContent: 'center',
-                                                            flexShrink: 0,
-                                                        }}
-                                                    >
-                                                        {result.type === 'tv' ? (
-                                                            <IcoTv size={14} color={sv.inkFaint} />
-                                                        ) : (
-                                                            <IcoMovie size={14} color={sv.inkFaint} />
-                                                        )}
-                                                    </div>
-                                                )}
-                                                <div style={{ minWidth: 0, flex: 1 }}>
-                                                    <p
-                                                        style={{
-                                                            fontFamily: sv.mono,
-                                                            fontSize: 13,
-                                                            color: sv.cyanHi,
-                                                            margin: 0,
-                                                            whiteSpace: 'nowrap',
-                                                            overflow: 'hidden',
-                                                            textOverflow: 'ellipsis',
-                                                        }}
-                                                    >
-                                                        {result.name}
-                                                    </p>
-                                                    <div style={{ display: 'flex', gap: 8, alignItems: 'center', marginTop: 2 }}>
-                                                        <span
-                                                            style={{
-                                                                fontFamily: sv.mono,
-                                                                fontSize: 10,
-                                                                textTransform: 'uppercase',
-                                                                padding: '2px 6px',
-                                                                color: result.type === 'tv' ? sv.cyan : sv.magenta,
-                                                                border: `1px solid ${result.type === 'tv' ? sv.cyan : sv.magenta}4d`,
-                                                                background: `${result.type === 'tv' ? sv.cyan : sv.magenta}1a`,
-                                                                letterSpacing: '0.14em',
-                                                            }}
-                                                        >
-                                                            {result.type}
-                                                        </span>
-                                                        {result.year && (
-                                                            <span style={{ fontFamily: sv.mono, fontSize: 10, color: sv.inkDim }}>
-                                                                {result.year}
-                                                            </span>
-                                                        )}
-                                                    </div>
-                                                </div>
-                                            </button>
-                                        ))}
-                                    </motion.div>
-                                )}
-                            </AnimatePresence>
-                        </div>
-
-                        <div style={{ height: 1, background: sv.line }} />
-
-                        {/* Title Input */}
-                        <div style={{ display: 'flex', flexDirection: 'column', gap: 8 }}>
-                            <SvLabel size={10}>Title</SvLabel>
-                            <input
-                                ref={titleInputRef}
-                                type="text"
-                                value={title}
-                                onChange={(e) => {
-                                    setTitle(e.target.value);
-                                    setTmdbId(undefined);
-                                    setSelectedYear(undefined);
-                                }}
-                                placeholder="e.g. Thunderbirds"
-                                style={inputStyle(!!title)}
-                                onFocus={(e) => (e.currentTarget.style.borderColor = sv.cyan)}
-                                onBlur={(e) =>
-                                    (e.currentTarget.style.borderColor = title ? sv.lineHi : sv.lineMid)
-                                }
-                            />
-                            {/* Confirms which TMDB show is selected (year + id) so the
-                                user isn't picking a same-name show blind. Only shown
-                                once a search result / candidate set the tmdbId. */}
-                            {tmdbId != null && (
-                                <span
-                                    style={{
-                                        fontFamily: sv.mono,
-                                        fontSize: 11,
-                                        color: sv.cyan,
-                                    }}
-                                >
-                                    Selected → {title}
-                                    {selectedYear ? ` (${selectedYear})` : ''} · TMDB #{tmdbId}
-                                </span>
-                            )}
-                        </div>
-
-                        {/* Media Type Toggle */}
-                        <div style={{ display: 'flex', flexDirection: 'column', gap: 8 }}>
-                            <SvLabel size={10}>Media Type</SvLabel>
-                            <div style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 8 }}>
-                                {(
-                                    [
-                                        { value: 'movie', label: 'Movie', Icon: IcoMovie },
-                                        { value: 'tv', label: 'TV Show', Icon: IcoTv },
-                                    ] as const
-                                ).map(({ value, label, Icon }) => {
-                                    const active = contentType === value;
-                                    return (
-                                        <motion.button
-                                            key={value}
-                                            type="button"
-                                            onClick={() => setContentType(value)}
-                                            whileHover={{ scale: 1.02 }}
-                                            whileTap={{ scale: 0.98 }}
-                                            style={{
-                                                display: 'flex',
-                                                alignItems: 'center',
-                                                justifyContent: 'center',
-                                                gap: 8,
-                                                padding: '10px 14px',
-                                                fontFamily: sv.mono,
-                                                fontSize: 11,
-                                                fontWeight: 700,
-                                                letterSpacing: '0.18em',
-                                                textTransform: 'uppercase',
-                                                color: active ? sv.cyanHi : sv.inkDim,
-                                                border: `1px solid ${active ? sv.cyan : sv.lineMid}`,
-                                                background: active ? `${sv.cyan}14` : 'transparent',
-                                                boxShadow: active
-                                                    ? `0 0 12px ${sv.cyan}4d, inset 0 0 8px ${sv.cyan}0d`
-                                                    : 'none',
-                                                cursor: 'pointer',
-                                                transition: 'all 0.18s',
-                                            }}
-                                        >
-                                            <Icon size={14} />
-                                            {label}
-                                        </motion.button>
-                                    );
-                                })}
-                            </div>
-                        </div>
-
-                        <AnimatePresence>
-                            {contentType === 'tv' && (
-                                <motion.div
-                                    initial={{ opacity: 0, height: 0 }}
-                                    animate={{ opacity: 1, height: 'auto' }}
-                                    exit={{ opacity: 0, height: 0 }}
-                                    transition={{ type: 'spring', stiffness: 400, damping: 35 }}
-                                    style={{ overflow: 'hidden' }}
-                                >
-                                    <div style={{ display: 'flex', flexDirection: 'column', gap: 8, paddingTop: 4 }}>
-                                        <SvLabel size={10}>Season</SvLabel>
-                                        <input
-                                            type="number"
-                                            min={1}
-                                            max={99}
-                                            value={season}
-                                            onChange={(e) => setSeason(e.target.value)}
-                                            style={{ ...inputStyle(true), width: 128, background: sv.bg0 }}
-                                            onFocus={(e) => (e.currentTarget.style.borderColor = sv.cyan)}
-                                            onBlur={(e) => (e.currentTarget.style.borderColor = sv.lineHi)}
-                                        />
-                                    </div>
-                                </motion.div>
-                            )}
-                        </AnimatePresence>
+                        <IdentityFields
+                            value={identity}
+                            onChange={setIdentity}
+                            onSearchQueryChange={(query) => setHasSearchQuery(!!query)}
+                            autoFocus
+                        />
 
                         <div style={{ height: 1, background: sv.line }} />
 
@@ -648,10 +334,10 @@ export default function ReIdentifyModal({ job, onSubmit, onCancel }: ReIdentifyM
                             <motion.button
                                 type="button"
                                 onClick={handleSubmit}
-                                disabled={!title.trim()}
+                                disabled={!identity.title.trim()}
                                 data-testid="reidentify-submit"
-                                whileHover={title.trim() ? { scale: 1.02 } : {}}
-                                whileTap={title.trim() ? { scale: 0.97 } : {}}
+                                whileHover={identity.title.trim() ? { scale: 1.02 } : {}}
+                                whileTap={identity.title.trim() ? { scale: 0.97 } : {}}
                                 style={{
                                     flex: 1,
                                     padding: '10px 16px',
@@ -660,14 +346,14 @@ export default function ReIdentifyModal({ job, onSubmit, onCancel }: ReIdentifyM
                                     fontWeight: 700,
                                     letterSpacing: '0.18em',
                                     textTransform: 'uppercase',
-                                    color: title.trim() ? sv.cyan : `${sv.cyan}4d`,
-                                    border: `1px solid ${title.trim() ? sv.cyan : `${sv.cyan}33`}`,
-                                    background: title.trim() ? `${sv.cyan}1f` : 'transparent',
-                                    boxShadow: title.trim()
+                                    color: identity.title.trim() ? sv.cyan : `${sv.cyan}4d`,
+                                    border: `1px solid ${identity.title.trim() ? sv.cyan : `${sv.cyan}33`}`,
+                                    background: identity.title.trim() ? `${sv.cyan}1f` : 'transparent',
+                                    boxShadow: identity.title.trim()
                                         ? `0 0 16px ${sv.cyan}4d, inset 0 0 8px ${sv.cyan}0d`
                                         : 'none',
-                                    cursor: title.trim() ? 'pointer' : 'not-allowed',
-                                    opacity: title.trim() ? 1 : 0.3,
+                                    cursor: identity.title.trim() ? 'pointer' : 'not-allowed',
+                                    opacity: identity.title.trim() ? 1 : 0.3,
                                 }}
                             >
                                 Re-Identify
